Launch our patched ghostdriver instead of the built-in edition
At least for the time being, ghostdriver/WebDriver do not want to
support returning the HTTP status of the response, and also not
the HTTP headers that the response returned. While this is a valid
decision in the greater picture of Selenium only intending to
automate the user input to a browser, it is useful to provide
additional capabilities where the browser provides them.